Changelog: http://itextpdf.com/changelog/132
Release Notes: http://itextpdf.com/release/iText132
Download: https://github.com/itext/itextpdf/releases/tag/1.3.2

+++ itext/src/com/lowagie/text/List.java
@@ -189,7 +189,7 @@ public class List implements TextElementArray, MarkupAttributes {
         }
         this.symbolIndent = 0;
         if ((value = (String)attributes.remove(ElementTags.SYMBOLINDENT)) != null) {
-            this.symbolIndent = Integer.parseInt(value);
+            this.symbolIndent = Float.parseFloat(value);
         }
         
         if ((value = (String)attributes.remove(ElementTags.FIRST)) != null) {
